fix: move non-react code to outside of pages directory
